Fix forms header bug with folder selector being rendered on submission headers
