+++ features/folders/view-forms-header/index.ts
@@ -1,8 +1,8 @@
 export { buildFormsBreadcrumbModel } from "./build-forms-breadcrumb-model.server";
 export { getFormsHeaderDataCached } from "./get-forms-header-data.server";
+export { FormsFolderHeaderSlot } from "./forms-folder-header-slot";
 export {
   folderSlugsMatch,
   normalizeFolderSlug,
 } from "@/features/folders/lib/folder-slug-matching";
-export { resolveFormsHeaderRouteContext } from "./resolve-forms-header-route-context";
 export { resolveFolderForNavBySlug } from "@/features/folders/get-folder-by-slug/resolve-folder-for-nav.server";
